Adjustments for Rod on coupons
Sorry I was not clear, only the 2 should be implemented.
1.       Populate PROMO_CODE with the promotion code like ‘MOREPOO’.
This is important so MOM can record what promotion code used. Paige will
create Promotion Code in MOM as needed.

2.       Populate PROMOCRED with the discount amount with a positive
value. Since Foxycart computes the order discount and probably give a
negative amount we need to convert it to a positive amount.

Please remove the other changes by doing the following:

1.       Remove the promotion code in PRODUCT01..05 columns. Example in
order id 151999201, remove PROMO-MOREPOO in PRODUCT02 and 1 in
QUANTITY02.

2.       For the TRYITFREE PP-5ML, this should not be affected with the
change since the change is only for the promotional codes entered by the
customers. Remove 5 in the PROMOCRED. Remove FreeSample in PROMO_CODE.

Please correct the following:

1.       In the same order 151999201, the PROMOCRED is 6 which is
incorrect. It should be 6.48.